requirements.txt
pip==22.2
setuptools==65.5
wheel==0.38

bidict
eclipse-sumo==1.8
libsumo==1.8
matplotlib
numpy==1.26.0
pandas
protobuf==3.20.1
pyyaml
ray==1.11
ray[rllib]==1.11
ray[tune]==1.11
scipy
sparse
sumolib==1.8
tqdm
traci==1.8
ujson
xmltodict
ipykernel
